requirements.old.txt
# Core Framework
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
sqlalchemy>=2.0.23
alembic>=1.13.1
ua-parser>=0.18.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
email-validator>=2.1.0.post1

# Database
psycopg2-binary>=2.9.9
redis>=5.0.1
pymysql>=1.1.0

# AI/ML Core
tensorflow>=2.16.1,<2.17.0
torch>=2.2.2
scikit-learn>=1.3.2
pandas>=2.1.4
numpy>=1.25.2

# Advanced ML
xgboost>=2.0.0
lightgbm>=4.0.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.8.5

# Graph Database
neo4j==5.14.0

# Explainability
shap==0.43.0
lime==0.2.0.1

# Data Quality & Drift Detection
evidently==0.4.14
great-expectations==0.18.8
alibi-detect==0.11.2

# AutoML
auto-sklearn==0.15.0
optuna==3.4.0
hyperopt==0.2.7

# Deep Learning
keras==2.15.0
transformers==4.36.2
sentence-transformers==2.2.2

# Time Series
prophet==1.1.4
statsmodels==0.14.0

# Privacy & Security
diffprivlib==0.6.0
federated-learning==1.0.0

# Streaming & Real-time
kafka-python==2.0.2
pyspark==3.5.0
streamz==0.6.4

# Model Serving
mlflow==2.8.1
bentoml>=1.4.26

# Security
cryptography==41.0.8
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.1.2

# Blockchain
web3==6.11.4
eth-account==0.9.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# API Integration
httpx==0.25.2

# Data Processing
pydantic-settings==2.1.0
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

